rename + auxiliary file deleted
