+++ crates/conjure_core/src/solver/adaptors/mod.rs
@@ -4,10 +4,10 @@
 // pub use kissat::Kissat;
 #[doc(inline)]
 pub use minion::Minion;
-pub use sat_adaptor::SAT;
+pub use rustsat::SAT;
 // mod sat_common;
 
 // mod kissat;
 mod minion;
-pub mod sat_adaptor;                                //temp visibility
+pub mod rustsat;                                //temp visibility
 // mod common;
