+++ crates/conjure_core/src/solver/adaptors/rustsat.rs
